# executor: remove redundant memory pre-allocations in parallel sort executor