@xzhangxian1008 2024-06-18
> ### What problem does this PR solve?
> Issue Number: close #54070
> 
> Problem Summary:
> 
> ### What changed and how does it work?
> ### Check List
> Tests
> 
> * [x]  Unit test
> * [ ]  Integration test
> * [ ]  Manual test (add detailed scripts or steps below)
> * [ ]  No need to test
>   > * [ ]  I checked and no code files have been changed.
> 
> Side effects
> 
> * [ ]  Performance regression: Consumes more CPU
> * [ ]  Performance regression: Consumes more Memory
> * [ ]  Breaking backward compatibility
> 
> Documentation
> 
> * [ ]  Affects user behaviors
> * [ ]  Contains syntax changes
> * [ ]  Contains variable changes
> * [ ]  Contains experimental features
> * [ ]  Changes MySQL compatibility
> 
> ### Release note
> Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.
> 
> ```
> Remove redundant memory pre-allocations in parallel sort executor
> ```



> Could you update the PR description by adding how the performance regression is introduced and how this fixes it?

updated