@windtalker 2024-06-20
How about we just save the orignal `chunk` in `addChunkToBatchRows`, and only construct the `batchRows` before `sortBatchRows`? So in `sortBatchRows` we already know the row size, and can pre-allocate it without waste.